[Refactor] Split HDFS scanner context from scanner runtime
