+++ be/src/formats/orc/orc_input_stream.cpp
@@ -15,6 +15,7 @@
 #include "formats/orc/orc_input_stream.h"
 
 #include "common/config_scan_io_fwd.h"
+#include "exec/hdfs_scanner/hdfs_scanner_context.h"
 #include "exprs/cast_expr.h"
 #include "formats/orc/orc_mapping.h"
 #include "fs/fs.h"
