+++ be/src/formats/parquet/metadata.h
@@ -17,12 +17,16 @@
 #include <string>
 
 #include "common/status.h"
-#include "exec/hdfs_scanner/hdfs_scanner.h"
+#include "exec/hdfs_scanner/hdfs_scanner_context.h"
 #include "formats/parquet/schema.h"
 #include "fs/fs.h"
 #include "gen_cpp/parquet_types.h"
 #include "storage/rowset/page_handle_fwd.h"
 #include "types/logical_type.h"
+
+namespace starrocks {
+class StoragePageCache;
+} // namespace starrocks
 
 namespace starrocks::parquet {
