+++ be/src/formats/parquet/page_reader.cpp
@@ -20,18 +20,17 @@
 #include <ostream>
 #include <vector>
 
+#include "base/compression/block_compression.h"
 #include "base/container/raw_container.h"
 #include "cache/datacache.h"
 #include "common/compiler_util.h"
 #include "common/config_scan_io_fwd.h"
 #include "common/status.h"
 #include "common/util/thrift_util.h"
-#include "exec/hdfs_scanner/hdfs_scanner.h"
 #include "formats/parquet/column_reader.h"
 #include "formats/parquet/utils.h"
 #include "gutil/strings/substitute.h"
 #include "runtime/current_thread.h"
-#include "runtime/exec_env.h"
 
 namespace starrocks::parquet {
